Started creating collection persist tests
