+++ packages/core/tests/new/collection/collection.persist.test.ts
@@ -1,223 +1,103 @@
-import { Agile, Item } from "../../../src";
+import { Agile, CollectionPersistent, Collection } from "../../../src";
 
-describe("Collection Persist Function Tests", () => {
-  const myStorage: any = {};
-  const storageMethods = {
-    get: jest.fn((key) => {
-      // console.log(`GET '${key}'`);
-      return myStorage[key];
-    }),
-    set: jest.fn((key, value) => {
-      // console.log(`SET '${key}'`, value);
-      myStorage[key] = value;
-    }),
-    remove: jest.fn((key) => {
-      // console.log(`DELETE '${key}'`);
-      delete myStorage[key];
-    }),
-  };
+describe("CollectionPersist Tests", () => {
+  let dummyAgile: Agile;
+  let dummyCollection: Collection;
 
-  // Define Agile with Storage
-  const App = new Agile();
-  App.registerStorage(
-    App.Storage({
-      key: "testStorage",
-      prefix: "test",
-      methods: storageMethods,
-    })
-  );
+  beforeEach(() => {
+    jest.clearAllMocks();
 
-  interface User {
-    id: number;
-    name: string;
-  }
+    dummyAgile = new Agile({ localStorage: false });
+    dummyCollection = new Collection(dummyAgile);
 
-  describe("Collection", () => {
-    it("Can persist Collection", async () => {
-      // Create Collection
-      const MY_COLLECTION = App.Collection<User>();
+    jest.spyOn(CollectionPersistent.prototype, "instantiatePersistent");
+    jest.spyOn(CollectionPersistent.prototype, "initialLoading");
+    console.error = jest.fn();
+  });
 
-      // Test Collecting Item before Persisting
-      MY_COLLECTION.collect({ id: 2, name: "hans" });
-      await new Promise((resolve) => setTimeout(resolve, 100));
-
-      expect(myStorage).toStrictEqual({});
-      expect(storageMethods.set).not.toHaveBeenCalled();
-      expect(storageMethods.get).not.toHaveBeenCalled();
-      expect(storageMethods.remove).not.toHaveBeenCalled();
-
-      // Test Persisting
-      MY_COLLECTION.persist("myCollection");
-      await new Promise((resolve) => setTimeout(resolve, 100));
-
-      expect(myStorage).toStrictEqual({
-        _test_myCollection: "true",
-        _test__myCollection_group_default: "[2]",
-        _test__myCollection_item_2: '{"id":2,"name":"hans"}',
+  it("should create CollectionPersistent and shouldn't call initialLoading if Persistent isn't ready (default config)", () => {
+    // Overwrite instantiatePersistent once to not call it and set ready property
+    jest
+      .spyOn(CollectionPersistent.prototype, "instantiatePersistent")
+      .mockImplementationOnce(function () {
+        this.ready = false;
       });
-      expect(storageMethods.set).toHaveBeenCalledTimes(3);
-      expect(storageMethods.get).toHaveBeenCalledTimes(3);
-      expect(storageMethods.remove).toHaveBeenCalledTimes(0);
 
-      // Test collecting new Item
-      MY_COLLECTION.collect({ id: 1, name: "frank" });
-      await new Promise((resolve) => setTimeout(resolve, 100));
+    const collectionPersistent = new CollectionPersistent(dummyCollection);
 
-      expect(myStorage).toStrictEqual({
-        _test_myCollection: "true",
-        _test__myCollection_group_default: "[2,1]",
-        _test__myCollection_item_2: '{"id":2,"name":"hans"}',
-        _test__myCollection_item_1: '{"id":1,"name":"frank"}',
+    expect(collectionPersistent).toBeInstanceOf(CollectionPersistent);
+    expect(collectionPersistent.collection()).toBe(dummyCollection);
+    expect(collectionPersistent.instantiatePersistent).toHaveBeenCalledWith({
+      key: undefined,
+      storageKeys: [],
+    });
+    expect(collectionPersistent.initialLoading).not.toHaveBeenCalled();
+
+    expect(collectionPersistent._key).toBe(CollectionPersistent.placeHolderKey);
+    expect(collectionPersistent.ready).toBeFalsy();
+    expect(collectionPersistent.isPersisted).toBeFalsy();
+    expect(collectionPersistent.onLoad).toBeUndefined();
+    expect(collectionPersistent.storageKeys).toStrictEqual([]);
+    expect(collectionPersistent.defaultStorageKey).toBeUndefined();
+  });
+
+  it("should create CollectionPersistent and shouldn't call initialLoading if Persistent isn't ready (specific config)", () => {
+    // Overwrite instantiatePersistent once to not call it and set ready property
+    jest
+      .spyOn(CollectionPersistent.prototype, "instantiatePersistent")
+      .mockImplementationOnce(function () {
+        this.ready = false;
       });
-      expect(storageMethods.set).toHaveBeenCalledTimes(5);
-      expect(storageMethods.get).toHaveBeenCalledTimes(4);
-      expect(storageMethods.remove).toHaveBeenCalledTimes(0);
 
-      // Test creating Group
-      MY_COLLECTION.createGroup("stuipidPeople", [1, 2]).persist({
-        followCollectionPersistKeyPattern: true,
-      });
-      await new Promise((resolve) => setTimeout(resolve, 100));
-
-      expect(myStorage).toStrictEqual({
-        _test_myCollection: "true",
-        _test__myCollection_group_stuipidPeople: "[1,2]",
-        _test__myCollection_group_default: "[2,1]",
-        _test__myCollection_item_2: '{"id":2,"name":"hans"}',
-        _test__myCollection_item_1: '{"id":1,"name":"frank"}',
-      });
-      expect(storageMethods.set).toHaveBeenCalledTimes(6);
-      expect(storageMethods.get).toHaveBeenCalledTimes(5);
-      expect(storageMethods.remove).toHaveBeenCalledTimes(0);
-
-      // Test collecting new Item
-      MY_COLLECTION.collect({ id: 3, name: "günter" });
-      await new Promise((resolve) => setTimeout(resolve, 100));
-
-      expect(myStorage).toStrictEqual({
-        _test_myCollection: "true",
-        _test__myCollection_group_stuipidPeople: "[1,2]",
-        _test__myCollection_group_default: "[2,1,3]",
-        _test__myCollection_item_2: '{"id":2,"name":"hans"}',
-        _test__myCollection_item_1: '{"id":1,"name":"frank"}',
-        _test__myCollection_item_3: '{"id":3,"name":"günter"}',
-      });
-      expect(storageMethods.set).toHaveBeenCalledTimes(8);
-      expect(storageMethods.get).toHaveBeenCalledTimes(6);
-      expect(storageMethods.remove).toHaveBeenCalledTimes(0);
-
-      // Test updating Item
-      MY_COLLECTION.update(3, { name: "Benno" });
-      await new Promise((resolve) => setTimeout(resolve, 100));
-
-      expect(myStorage).toStrictEqual({
-        _test_myCollection: "true",
-        _test__myCollection_group_stuipidPeople: "[1,2]",
-        _test__myCollection_group_default: "[2,1,3]",
-        _test__myCollection_item_2: '{"id":2,"name":"hans"}',
-        _test__myCollection_item_1: '{"id":1,"name":"frank"}',
-        _test__myCollection_item_3: '{"id":3,"name":"Benno"}',
-      });
-      expect(storageMethods.set).toHaveBeenCalledTimes(9);
-      expect(storageMethods.get).toHaveBeenCalledTimes(6);
-      expect(storageMethods.remove).toHaveBeenCalledTimes(0);
-
-      // Test updating Item with ItemKey
-      MY_COLLECTION.update(1, { id: 37, name: "Arne" });
-      await new Promise((resolve) => setTimeout(resolve, 100));
-
-      expect(myStorage).toStrictEqual({
-        _test_myCollection: "true",
-        _test__myCollection_group_stuipidPeople: "[37,2]",
-        _test__myCollection_group_default: "[2,37,3]",
-        _test__myCollection_item_2: '{"id":2,"name":"hans"}',
-        _test__myCollection_item_37: '{"id":37,"name":"Arne"}',
-        _test__myCollection_item_3: '{"id":3,"name":"Benno"}',
-      });
-      expect(storageMethods.set).toHaveBeenCalledTimes(12);
-      expect(storageMethods.get).toHaveBeenCalledTimes(6);
-      expect(storageMethods.remove).toHaveBeenCalledTimes(1);
+    const collectionPersistent = new CollectionPersistent(dummyCollection, {
+      key: "collectionPersistentKey",
+      storageKeys: ["test1", "test2"],
     });
 
-    it("Can load persisted Collection", async () => {
-      // Create Collection
-      const MY_COLLECTION = App.Collection<User>();
+    expect(collectionPersistent).toBeInstanceOf(CollectionPersistent);
+    expect(collectionPersistent.instantiatePersistent).toHaveBeenCalledWith({
+      key: "collectionPersistentKey",
+      storageKeys: ["test1", "test2"],
+    });
+    expect(collectionPersistent.initialLoading).not.toHaveBeenCalled();
 
-      // Load persisted Value
-      MY_COLLECTION.persist("myCollection");
-      await new Promise((resolve) => setTimeout(resolve, 100));
+    expect(collectionPersistent._key).toBe(CollectionPersistent.placeHolderKey);
+    expect(collectionPersistent.ready).toBeFalsy();
+    expect(collectionPersistent.isPersisted).toBeFalsy();
+    expect(collectionPersistent.onLoad).toBeUndefined();
+    expect(collectionPersistent.storageKeys).toStrictEqual([]);
+    expect(collectionPersistent.defaultStorageKey).toBeUndefined();
+  });
 
-      expect(myStorage).toStrictEqual({
-        _test_myCollection: "true",
-        _test__myCollection_group_stuipidPeople: "[37,2]",
-        _test__myCollection_group_default: "[2,37,3]",
-        _test__myCollection_item_2: '{"id":2,"name":"hans"}',
-        _test__myCollection_item_37: '{"id":37,"name":"Arne"}',
-        _test__myCollection_item_3: '{"id":3,"name":"Benno"}',
-      });
-      expect(MY_COLLECTION.isPersisted).toBeTruthy();
-      expect(MY_COLLECTION.size).toBe(3);
-      expect(MY_COLLECTION.data["2"]).toBeInstanceOf(Item);
-      expect(MY_COLLECTION.data["37"]).toBeInstanceOf(Item);
-      expect(MY_COLLECTION.data["3"]).toBeInstanceOf(Item);
-
-      // Updating some Collection Stuff
-      MY_COLLECTION.update(3, { name: "Angela" });
-      MY_COLLECTION.collect({ id: 4, name: "Paul" });
-      MY_COLLECTION.collect({ id: 99, name: "Jeff" });
-      await new Promise((resolve) => setTimeout(resolve, 100));
-
-      expect(myStorage).toStrictEqual({
-        _test_myCollection: "true",
-        _test__myCollection_group_default: "[2,37,3,4,99]",
-        _test__myCollection_item_2: '{"id":2,"name":"hans"}',
-        _test__myCollection_group_stuipidPeople: "[37,2]",
-        _test__myCollection_item_3: '{"id":3,"name":"Angela"}',
-        _test__myCollection_item_37: '{"id":37,"name":"Arne"}',
-        _test__myCollection_item_4: '{"id":4,"name":"Paul"}',
-        _test__myCollection_item_99: '{"id":99,"name":"Jeff"}',
+  it("should create CollectionPersistent and should call initialLoading if Persistent is ready (default config)", () => {
+    // Overwrite instantiatePersistent once to not call it
+    jest
+      .spyOn(CollectionPersistent.prototype, "instantiatePersistent")
+      .mockImplementationOnce(function () {
+        this.ready = true;
       });
 
-      // Test removing Item
-      MY_COLLECTION.remove(3).everywhere();
-      await new Promise((resolve) => setTimeout(resolve, 100));
+    const collectionPersistent = new CollectionPersistent(dummyCollection);
 
-      expect(myStorage).toStrictEqual({
-        _test_myCollection: "true",
-        _test__myCollection_group_default: "[2,37,4,99]",
-        _test__myCollection_item_2: '{"id":2,"name":"hans"}',
-        _test__myCollection_group_stuipidPeople: "[37,2]",
-        _test__myCollection_item_37: '{"id":37,"name":"Arne"}',
-        _test__myCollection_item_4: '{"id":4,"name":"Paul"}',
-        _test__myCollection_item_99: '{"id":99,"name":"Jeff"}',
+    expect(collectionPersistent.initialLoading).toHaveBeenCalled();
+  });
+
+  it("should create CollectionPersistent and shouldn't call initialLoading if Persistent is ready (config.instantiate = false)", () => {
+    // Overwrite instantiatePersistent once to not call it and set ready property
+    jest
+      .spyOn(CollectionPersistent.prototype, "instantiatePersistent")
+      .mockImplementationOnce(function () {
+        this.ready = true;
       });
+
+    const collectionPersistent = new CollectionPersistent(dummyCollection, {
+      instantiate: false,
     });
 
-    it("Can remove persisted Collection", async () => {
-      // Create Collection
-      const MY_COLLECTION = App.Collection<User>();
+    expect(collectionPersistent.initialLoading).not.toHaveBeenCalled();
+  });
 
-      // Load persisted Value
-      MY_COLLECTION.persist("myCollection");
-      await new Promise((resolve) => setTimeout(resolve, 100));
-
-      expect(myStorage).toStrictEqual({
-        _test_myCollection: "true",
-        _test__myCollection_group_default: "[2,37,4,99]",
-        _test__myCollection_item_2: '{"id":2,"name":"hans"}',
-        _test__myCollection_group_stuipidPeople: "[37,2]",
-        _test__myCollection_item_37: '{"id":37,"name":"Arne"}',
-        _test__myCollection_item_4: '{"id":4,"name":"Paul"}',
-        _test__myCollection_item_99: '{"id":99,"name":"Jeff"}',
-      });
-
-      // Test Removing Persisted Value
-      MY_COLLECTION.persistent?.removePersistedValue();
-      await new Promise((resolve) => setTimeout(resolve, 100));
-
-      expect(myStorage).toStrictEqual({
-        _test__myCollection_group_stuipidPeople: "[37,2]",
-      });
-    });
+  describe("CollectionPersistent Function Tests", () => {
+    // TODO
   });
 });
